refactor and seprate the fiels
