--- lib/features/community/ui/widgets/posts_list.dart
@@ -1,21 +1,0 @@
-import 'package:tranquilo_app/features/community/data/model/post.dart';
-
-final List<Post> posts = [
-  Post(
-    author: "Anonymous",
-    date: "15 August on 4:22 AM",
-    content:
-        "I’ve been practicing mindfulness and grounding exercises for a while...",
-    upvotes: 10,
-    downvotes: 0,
-    comments: 8,
-  ),
-  Post(
-    author: "Yara Ahmed",
-    date: "1 August on 11:19 PM",
-    content: "Let’s check in with each other! How are you feeling today...",
-    upvotes: 9,
-    downvotes: 1,
-    comments: 6,
-  ),
-];
